requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-openai==0.2.1
langsmith==0.1.128
openai==1.48.0
faiss-cpu==1.8.0.post1
PyPDF2==3.0.1
mysql-connector==2.2.9
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
tiktoken==0.7.0

